+++ app/views/condition/_condition_list.html.haml
@@ -1,0 +1,21 @@
+#condition_list_div
+  - if @conditions
+    %div{:style => "padding-top: 10px;"}
+    = render :partial => "layouts/flash_msg"
+    - if @conditions.empty?
+      - model = ui_lookup(:model => @sb[:folder].camelize)
+      - if @search_text.blank?
+        - msg = _("No %{model} Conditions are defined.") % {:model => model}
+      - else
+        - msg = _("No %{model} Conditions are defined that match the entered search string.") % {:model => model}
+      = render :partial => 'layouts/info_msg', :locals => {:message => msg}
+    - else
+      %table.table.table-striped.table-bordered.table-hover
+        %tbody
+          - @conditions.each do |c|
+            %tr{:title => _("View this Condition"),
+              :onclick => "miqTreeActivateNode('condition_tree', 'xx-#{c.towhat.camelize(:lower)}_co-#{c.id}');"}
+              %td.table-view-pf-select
+                %i.fa.fa-diamond
+              %td
+                = c.description
