+++ app/views/condition/explorer.html.haml
@@ -1,0 +1,46 @@
+- content_for :search do
+  = render(:partial => "layouts/x_adv_searchbox", :locals => {:nameonly => true})
+
+#main_div
+  - case @nodetype
+  - when "root"
+    - case x_active_tree
+    - when :policy_profile_tree
+      = render :partial => "profile_list"
+    - when :policy_tree
+      = render :partial => "policy_folders"
+    - when :event_tree
+      = render :partial => "event_list"
+    - when :condition_tree
+      = render :partial => "condition_folders"
+    - when :action_tree
+      = render :partial => "action_list"
+    - when :alert_profile_tree
+      = render :partial => "alert_profile_folders"
+    - when :alert_tree
+      = render :partial => "alert_list"
+  - when "pp"
+    = render :partial => "profile_details"
+  - when "xx"
+    - if @profiles
+      = render :partial => "profile_list"
+    - elsif @policies || (@view && @sb[:tree_typ] == "policies")
+      = render :partial => "policy_list"
+    - elsif @conditions
+      = render :partial => "condition_list"
+    - elsif @folders
+      = render :partial => "policy_folders"
+    - elsif @alert_profiles
+      = render :partial => "alert_profile_list"
+  - when "p"
+    = render :partial => "policy_details", :locals => {:read_only => true}
+  - when "co"
+    = render :partial => "condition_details", :locals => {:read_only => true}
+  - when "ev"
+    = render :partial => "event_details", :locals => {:read_only => true}
+  - when "a", "ta", "fa"
+    = render :partial => "action_details", :locals => {:read_only => true}
+  - when "ap"
+    = render :partial => "alert_profile_details"
+  - when "al"
+    = render :partial => "alert_details", :locals => {:read_only => true}
